requirements.txt
alabaster==0.7.10
amqp==2.2.2
Babel==2.5.3
backcall==0.1.0
beautifulsoup4==4.6.0
billiard==3.5.0.3
bokeh==0.12.16
celery==4.1.1
certifi==2018.4.16
chardet==3.0.4
decorator==4.3.0
Django==2.0.5
django-crispy-forms==1.7.2
django-formtools==2.1
django-htmlmin==0.10.0
django-sortedm2m==1.5.0
django-tinymce==2.7.0
docutils==0.14
elasticsearch==6.2.0
html5lib==1.0.1
idna==2.6
imagesize==1.0.0
ipdb==0.11
ipython==6.4.0
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.12.0
Jinja2==2.10
kombu==4.2.0
MarkupSafe==1.0
natsort==5.3.2
numpy==1.14.3
packaging==17.1
pandas==0.23.0
parso==0.2.1
pexpect==4.5.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pycodestyle==2.4.0
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.3
python-memcached==1.59
pytz==2018.4
PyYAML==3.12
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.4
sphinxcontrib-websupport==1.0.1
tornado==5.0.2
tqdm==4.23.3
traitlets==4.3.2
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
webencodings==0.5.1
